Backend finalizado

+++ backend/controllers/productControllers.js
@@ -3,7 +3,7 @@ const Product = require('../model/productModel')
 const { findByIdAndDelete } = require('../model/productModel')
 
 const getProduct = asyncHandler ( async (req, res) => {
-    const products = await Product.find()
+    const products = await Product.find({ user: req.user.id })
 
     res.status(200).json({ products })
 })
@@ -17,7 +17,8 @@ const createProduct = asyncHandler ( async (req, res) => {
     const product = await Product.create({
         nombre: req.body.nombre,
         precio: req.body.precio,
-        descripcion: req.body.descripcion
+        descripcion: req.body.descripcion,
+        user: req.user.id
     })
 
     res.status(201).json({ product })
@@ -26,17 +27,25 @@ const createProduct = asyncHandler ( async (req, res) => {
 const updateProduct = asyncHandler ( async (req, res) => {
     const product = await Product.findById(req.params.id)
 
+    //se verifica que el producto existe
     if(!product){
         res.status(400)
         throw new Error('El producto no fue encontrado')
     }
 
-    const productUpdated = await Product.findByIdAndUpdate(req.params.id, req.body, { new: true })
+    //se verifica que el producto pertenezca al usuario del token que lo quiera modificar
+    if(product.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    } else {
+        const productUpdated = await Product.findByIdAndUpdate(req.params.id, req.body, { new: true })
 
-    res.status(200).json(productUpdated)
+        res.status(200).json(productUpdated)
+    }
 })
 
 const deleteProduct = asyncHandler ( async (req, res) => {
+    //se verifica que la tarea exista
     const product = await Product.findById(req.params.id)
 
     if(!product){
@@ -44,10 +53,16 @@ const deleteProduct = asyncHandler ( async (req, res) => {
         throw new Error('El producto no fue encontrado')
     }
 
-    await Product.deleteOne(product)
-    //await Product.findByIdAndDelete(req.params.id)
+    //se verifica que el producto pertenezca al usuario del token que lo quiera modificar
+    if(product.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error('Acceso no autorizado')
+    } else {
+        await Product.deleteOne(product)
+        //await Product.findByIdAndDelete(req.params.id)
 
-    res.status(200).json({ id: req.params.id })
+        res.status(200).json({ id: req.params.id })
+    }
 })
 
 module.exports = {
